Implement Forth conditions and comments
Conditions and comments works by overriding the execution loop.

Comments override the execution loop starting on ( and until ) where
they restore the original loop.

Conditions override the execution loop on IF. At that point the top
element is popped and if it is true, the execution continues normally
until ELSE and then execution skip to after THEN. If the element is
false, execution is skipped until ELSE and THEN finishes the condition.

Code example are as follows:

    ( This is a comment )
    ( Don't forget the space after the brace )
    ( And this is a condition )
    1 < 2 IF TRUE_CONDITION ELSE FALSE_CONDITION THEN

The only problem is that we haven't implemented comparison operators.
That's why the tests fails.